requirements.txt
langchain==0.3.20
langchain-core==0.3.43
langchain-community==0.3.19
langchain-ollama==0.2.3
pypdf==5.3.1
pyyaml==6.0.2